Split task objects.
